// apps/backend/src/module/shared/model/entity/cfg/doattrs-all.entity.ts
/**
 * 一张视图表
 */
export interface DoattrsAllEntity {
  do_id: number

  ln_class: string

  do_name: string

  desc_cn: string

  unit: string

  ratio: number

  import_level: number

  do_order: number

  short_addr: number

  cdc_type_id: number

  precision: number

  cdc_type_desc: string
}